renamed pages for easier overview in our repo
