+++ frontend/src/pages/collaboration_graph.js
