+++ frontend/src/pages/trend_graphs.js
