+++ frontend/src/pages/world_map.js
